+++ src/compas/data/data.py
@@ -83,8 +83,8 @@ class Data(object):
     ...         self.y = data['y']
     ...         self.z = data['z']
     ...
-    >>> a = Point(1, 0, 0)
-    >>> a.guid
+    >>> a = Point(1.0, 0.0, 0.0)
+    >>> a.guid                 # doctest: +SKIP
     UUID('1ddad2fe-6716-4e30-a5ae-8ed7cad892c4')
     >>> a.name
     'Point'
@@ -93,10 +93,10 @@ class Data(object):
 
     >>> from compas.data import json_dumps, json_loads
     >>> s = json_dumps(a)
-    >>> b = json_loads(s)
-    >>> a is b
+    >>> b = json_loads(s)                                     # doctest: +SKIP
+    >>> a is b                                                # doctest: +SKIP
     False
-    >>> a == b
+    >>> a == b                                                # doctest: +SKIP
     True
 
     """