+++ src/compas/geometry/booleans/__init__.py
@@ -30,8 +30,8 @@ def boolean_union_mesh_mesh(A, B):
     Examples
     --------
     >>> from compas.geometry import Box, Sphere
-    >>> from compas.geometry import boolean_union_mesh_mesh
-    >>> from compas.geometry import trimesh_remesh
+    >>> from compas.geometry import boolean_union_mesh_mesh     # doctest: +SKIP
+    >>> from compas.geometry import trimesh_remesh              # doctest: +SKIP
     >>> from compas.datastructures import Mesh
 
     >>> box = Box.from_width_height_depth(2, 2, 2)
@@ -44,10 +44,10 @@ def boolean_union_mesh_mesh(A, B):
 
     >>> A = box.to_vertices_and_faces()
     >>> B = sphere.to_vertices_and_faces()
-    >>> B = trimesh_remesh(B, 0.3, 10)
+    >>> B = trimesh_remesh(B, 0.3, 10)                          # doctest: +SKIP
 
-    >>> V, F = boolean_union_mesh_mesh(A, B)
-    >>> union = Mesh.from_vertices_and_faces(V, F)
+    >>> V, F = boolean_union_mesh_mesh(A, B)                    # doctest: +SKIP
+    >>> union = Mesh.from_vertices_and_faces(V, F)              # doctest: +SKIP
 
     """
     raise NotImplementedError